Very radical change

+++ Owlyshield/owlyshield_predict/src/behavioral/rule_types.rs
@@ -1831,7 +1831,7 @@ pub struct NamedConditionGroup {
     #[serde(default, alias = "opsets", alias = "op_set", alias = "operation_sets")]
     pub irp_operations: Vec<String>,
     #[serde(default, alias = "opcodes", alias = "irp_ops")]
-    pub irp_opcodes: Vec<u8>,
+    pub irp_opcodes: Vec<u32>,
     #[serde(default)]
     pub pipe_payloads: Vec<ContentMatchData>,
     #[serde(default)]
